Add support for testing __STDC_WANT_IEC_60559_TYPES_EXT__
This is defined by TS 18661-3 for supporting the _FloatN
and _FloatNx types.

This will also implicitly enforce a minimum GCC 4.9.  Given
the standard is a C11 derivative, it would seem reasonable
to assume a C11 compiler with _Generic support.  This allows
for much simpler type classification macros which in theory
should be supported on any C11 compiler.

	* bits/libc-header-start.h:
	(__GLIBC_USR_IEC_60559_TYPES_EXT): New macro.