Fix arg used as litteral suffix in tst-strfrom.h
The macro ENTRY in tst-strfrom.h is used to generate the input values for
each floating-point type (float, double, long double).  It should append
the parameter LSUF (Literal suffix) to the floating-point number, but is
using CSUF (C function suffix).  This patch fixes it.

Tested for powerpc64le and x86_64.

2016-10-28  Gabriel F. T. Gomes  <gftg@linux.vnet.ibm.com>

	* stdlib/tst-strfrom.h (ENTRY): Replace use of CSUF with LSUF.